% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color.R
\name{scale_color_ifar}
\alias{scale_color_ifar}
\title{IFAR scale color}
\usage{
scale_color_ifar(..., discrete = FALSE, direction = 1,
  na.value = "grey50", guide = "colourbar")
}
\description{
IFAR scale color
}
