% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color.R
\name{scale_fill_ifar}
\alias{scale_fill_ifar}
\title{IFAR scale fill}
\usage{
scale_fill_ifar(..., discrete = FALSE, direction = 1, na.value = "grey50",
  guide = "colourbar")
}
\description{
IFAR scale fill
}
