% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_razor.R
\name{theme_razor}
\alias{theme_razor}
\title{Razor: Another minimalist ggplot theme}
\usage{
theme_razor(base_size = 11, base_family = "", stripe_background = FALSE,
  axis_y_title = FALSE)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{stripe_background}{T/F for whether facet stripes should be colored.
It also take color codes.Default is \code{F}.}

\item{axis_y_title}{T/F for the display of y axis title as the y axis title
is usually displayed through other mechanism in this design. Default is \code{F}.}
}
\description{
Razor: Another minimalist ggplot theme
}
